broke up edit dialog. if dropoff or pickup, customer appears.
